Refactor site relations table. #206
